Convert nearly all js files to ts
