--- lint-staged.config.js
@@ -1,7 +1,0 @@
-/**
- * @filename: lint-staged.config.js
- * @type {import('lint-staged').Configuration}
- */
-export default {
-  "*.{js,jsx,json,md,mjs,ts,tsx,yml,yaml}": "prettier --write",
-};
