+++ lint-staged.config.ts
@@ -1,0 +1,7 @@
+import type { Configuration } from "lint-staged";
+
+const config: Configuration = {
+  "*.{js,jsx,json,md,mjs,ts,tsx,yml,yaml}": "prettier --write",
+};
+
+export default config;
